+++ pkg/bindinfo/binding_operator_test.go
