+++ pkg/bindinfo/global_handle.go
@@ -15,19 +15,13 @@
 package bindinfo
 
 import (
-	"fmt"
 	"strings"
 	"time"
 
-	"github.com/pingcap/tidb/pkg/parser/ast"
-	"github.com/pingcap/tidb/pkg/parser/format"
-	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/sessionctx/vardef"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
 	"github.com/pingcap/tidb/pkg/util"
 	"github.com/pingcap/tidb/pkg/util/chunk"
-	"github.com/pingcap/tidb/pkg/util/hint"
-	"go.uber.org/zap"
 )
 
 // GlobalBindingHandle is used to handle all global sql bind operations.
@@ -77,18 +71,6 @@ func NewGlobalBindingHandle(sPool util.DestroyableSessionPool) GlobalBindingHand
 	return h
 }
 
-// lockBindInfoTable simulates `LOCK TABLE mysql.bind_info WRITE` by acquiring a pessimistic lock on a
-// special builtin row of mysql.bind_info. Note that this function must be called with h.sctx.Lock() held.
-// We can replace this implementation to normal `LOCK TABLE mysql.bind_info WRITE` if that feature is
-// generally available later.
-// This lock would enforce the CREATE / DROP GLOBAL BINDING statements to be executed sequentially,
-// even if they come from different tidb instances.
-func lockBindInfoTable(sctx sessionctx.Context) error {
-	// h.sctx already locked.
-	_, err := exec(sctx, LockBindInfoSQL)
-	return err
-}
-
 // newBindingFromStorage builds Bindings from a tuple in storage.
 func newBindingFromStorage(row chunk.Row) *Binding {
 	status := row.GetString(3)
@@ -109,64 +91,6 @@ func newBindingFromStorage(row chunk.Row) *Binding {
 		SQLDigest:   row.GetString(9),
 		PlanDigest:  row.GetString(10),
 	}
-}
-
-// GenerateBindingSQL generates binding sqls from stmt node and plan hints.
-func GenerateBindingSQL(stmtNode ast.StmtNode, planHint string, defaultDB string) string {
-	// If would be nil for very simple cases such as point get, we do not need to evolve for them.
-	if planHint == "" {
-		return ""
-	}
-	// We need to evolve plan based on the current sql, not the original sql which may have different parameters.
-	// So here we would remove the hint and inject the current best plan hint.
-	hint.BindHint(stmtNode, &hint.HintsSet{})
-	bindSQL := RestoreDBForBinding(stmtNode, defaultDB)
-	if bindSQL == "" {
-		return ""
-	}
-	switch n := stmtNode.(type) {
-	case *ast.DeleteStmt:
-		deleteIdx := strings.Index(bindSQL, "DELETE")
-		// Remove possible `explain` prefix.
-		bindSQL = bindSQL[deleteIdx:]
-		return strings.Replace(bindSQL, "DELETE", fmt.Sprintf("DELETE /*+ %s*/", planHint), 1)
-	case *ast.UpdateStmt:
-		updateIdx := strings.Index(bindSQL, "UPDATE")
-		// Remove possible `explain` prefix.
-		bindSQL = bindSQL[updateIdx:]
-		return strings.Replace(bindSQL, "UPDATE", fmt.Sprintf("UPDATE /*+ %s*/", planHint), 1)
-	case *ast.SelectStmt:
-		var selectIdx int
-		if n.With != nil {
-			var withSb strings.Builder
-			withIdx := strings.Index(bindSQL, "WITH")
-			restoreCtx := format.NewRestoreCtx(format.RestoreStringSingleQuotes|format.RestoreSpacesAroundBinaryOperation|format.RestoreStringWithoutCharset|format.RestoreNameBackQuotes, &withSb)
-			restoreCtx.DefaultDB = defaultDB
-			if err := n.With.Restore(restoreCtx); err != nil {
-				bindingLogger().Debug("restore SQL failed", zap.Error(err))
-				return ""
-			}
-			withEnd := withIdx + len(withSb.String())
-			tmp := strings.Replace(bindSQL[withEnd:], "SELECT", fmt.Sprintf("SELECT /*+ %s*/", planHint), 1)
-			return strings.Join([]string{bindSQL[withIdx:withEnd], tmp}, "")
-		}
-		selectIdx = strings.Index(bindSQL, "SELECT")
-		// Remove possible `explain` prefix.
-		bindSQL = bindSQL[selectIdx:]
-		return strings.Replace(bindSQL, "SELECT", fmt.Sprintf("SELECT /*+ %s*/", planHint), 1)
-	case *ast.InsertStmt:
-		insertIdx := int(0)
-		if n.IsReplace {
-			insertIdx = strings.Index(bindSQL, "REPLACE")
-		} else {
-			insertIdx = strings.Index(bindSQL, "INSERT")
-		}
-		// Remove possible `explain` prefix.
-		bindSQL = bindSQL[insertIdx:]
-		return strings.Replace(bindSQL, "SELECT", fmt.Sprintf("SELECT /*+ %s*/", planHint), 1)
-	}
-	bindingLogger().Debug("unexpected statement type when generating bind SQL", zap.Any("statement", stmtNode))
-	return ""
 }
 
 var (
